order_creator_ms/requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
psycopg2-binary==2.9.1
faker==9.3.1
pika==1.2.0
redis==5.2.1
PyJWT==2.10.1
requests==2.32.3
cryptography==44.0.1